Added example datasets, minor bugfix, updated data input
Example datasets were added to the repo and the Readme file was updated with instructions how to use them.
Fixed a bug in the aperture selection loop that was accidentally introduced with the last commit.
Modified the lcfit_utils.read_lc() function to ingest one ZP error column per dataset.